# perf(tree): avoid recursive calls