Reduce tree build time and improve module loading time.

Script from #2517

```
benchmark      time (avg)             (min … max)       p75       p99      p999
------------------------------------------------- -----------------------------
• built
------------------------------------------------- -----------------------------
current     76.94 µs/iter    (58.1 µs … 1,168 µs)   73.3 µs    279 µs  503.4 µs
new         59.77 µs/iter    (48.8 µs … 570.2 µs)   56.9 µs  197.8 µs    384 µs

summary for built
  new
   1.29x faster than current
```